examples/requirements.txt
# assume torch is already installed

# pycolmap for data parsing
 git+https://github.com/KohsukeIde/pycolmap.git@2b617eee014dc0ae7501afb4f9fb2b49ad238c6d
# (optional) nerfacc for torch version rasterization 
# git+https://github.com/nerfstudio-project/nerfacc

viser==0.2.7
nerfview==0.0.3
imageio[ffmpeg]
numpy<2.0.0
scikit-learn
tqdm
torchmetrics[image]
opencv-python
tyro>=0.8.8
Pillow
tensorboard
tensorly
pyyaml
matplotlib
git+https://github.com/rahul-goel/fused-ssim@1272e21a282342e89537159e4bad508b19b34157


lpips
pytorch-msssim
